requirements.txt
twilio
numpy
websocket_client
python-binance
TA-Lib
the Relative Strength Index (RSI) and Moving Average Convergence/Divergence (MACD)